refactor(#1188): separate regex string generator into individual classes
