+++ generator/src/main/java/com/scottlogic/deg/generator/generation/string/factorys/RandomStringFactory.java
@@ -1,5 +1,22 @@
-package com.scottlogic.deg.generator.generation.string;
+/*
+ * Copyright 2019 Scott Logic Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
+package com.scottlogic.deg.generator.generation.string.factorys;
+
+import com.scottlogic.deg.generator.generation.string.StringUtils;
 import com.scottlogic.deg.generator.utils.RandomNumberGenerator;
 import dk.brics.automaton.State;
 import dk.brics.automaton.Transition;
@@ -8,7 +25,7 @@ import java.util.List;
 
 public class RandomStringFactory {
 
-    String createRandomString(
+    public String createRandomString(
             String currentString,
             State state,
             int minLength,
