+++ generator/src/test/java/com/scottlogic/deg/generator/generation/string/AutomatonUtilsTests.java
@@ -31,144 +31,6 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class AutomatonUtilsTests {
-    
-    @Test
-    public void getLongestExample_withFixedLengthRegex_shouldReturnCorrectString(){
-        Automaton automaton = getAutomaton("[a-z]{9}");
-
-        String longestExample = AutomatonUtils.getLongestExample(automaton);
-
-        String expected = repeat('a', 9);
-        assertThat(longestExample, equalTo(expected));
-    }
-
-    @Test
-    public void getLongestExample_withRangeRegex_shouldReturnCorrectString(){
-        Automaton automaton = getAutomaton("[a-z]{0,9}");
-
-        String longestExample = AutomatonUtils.getLongestExample(automaton);
-
-        String expected = repeat('a', 9);
-        assertThat(longestExample, equalTo(expected));
-    }
-
-    @Test
-    public void getLongestExample_withLongRangeRegex_shouldReturnCorrectString(){
-        Automaton automaton = getAutomaton("[a-z]{0,1000}");
-
-        String longestExample = AutomatonUtils.getLongestExample(automaton);
-
-        String expected = repeat('a', 1000);
-        assertThat(longestExample, equalTo(expected));
-    }
-
-    @Test
-    public void getLongestExample_withSimpleRegexContainingOptionalPaths_shouldReturnCorrectString(){
-        Automaton automaton = getAutomaton("[f]{0,10}([a-c]{3}|[d-e]{5})");
-
-        String longestExample = AutomatonUtils.getLongestExample(automaton);
-
-        String expected = repeat('f', 10) + repeat('d', 5);
-        assertThat(longestExample, equalTo(expected));
-    }
-
-    @Test
-    public void getLongestExample_withComplexRegexContainingOptionalPaths_shouldReturnCorrectString(){
-        Automaton automaton = getAutomaton("[a-z]{0,10}([a-c]{3}|[d-e]{5})");
-
-        String longestExample = AutomatonUtils.getLongestExample(automaton);
-
-        assertThat(longestExample, matchesPattern("[a-z]{10}[d-e]{5}"));
-    }
-
-    @Test
-    public void getLongestExample_withLengthLimitedEmailAddress_shouldReturnCorrectStringWithLowestOrdinalCharPerSection(){
-        Automaton email = getAutomaton("^[a-z0-9]+\\@[a-z0-9]+\\.co(m|\\.uk)$");
-        Automaton lengthRestriction = getAutomaton("^.{0,20}$");
-        Automaton intersected = email.intersection(lengthRestriction);
-
-        String longestExample = AutomatonUtils.getLongestExample(intersected);
-
-        assertThat(longestExample.length(), equalTo(20));
-        assertThat(longestExample, matchesPattern("^[a-z0-9]+\\@[a-z0-9]+\\.co(m|\\.uk)$"));
-    }
-
-    @Test
-    public void getLongestExample_withEmptyString_shouldReturnEmptyString(){
-        Automaton automaton = getAutomaton("[a-z]{0}");
-
-        String longestExample = AutomatonUtils.getLongestExample(automaton);
-
-        assertThat(longestExample, equalTo(""));
-    }
-
-    @Test
-    public void getShortestExample_withFixedLengthRegex_shouldReturnCorrectString(){
-        Automaton automaton = getAutomaton("[a-z]{9}");
-
-        String shortestExample = AutomatonUtils.getShortestExample(automaton);
-
-        String expected = repeat('a', 9);
-        assertThat(shortestExample, equalTo(expected));
-    }
-
-    @Test
-    public void getShortestExample_withRangeRegex_shouldReturnCorrectString(){
-        Automaton automaton = getAutomaton("[a-z]{0,9}");
-
-        String shortestExample = AutomatonUtils.getShortestExample(automaton);
-
-        assertThat(shortestExample, equalTo(""));
-    }
-
-    @Test
-    public void getShortestExample_withLongRangeRegex_shouldReturnCorrectString(){
-        Automaton automaton = getAutomaton("[a-z]{0,1000}");
-
-        String shortestExample = AutomatonUtils.getShortestExample(automaton);
-
-        assertThat(shortestExample, equalTo(""));
-    }
-
-    @Test
-    public void getShortestExample_withSimpleRegexContainingOptionalPaths_shouldReturnCorrectString(){
-        Automaton automaton = getAutomaton("[f]{0,10}([a-c]{3}|[d-e]{5})");
-
-        String shortestExample = AutomatonUtils.getShortestExample(automaton);
-
-        String expected = repeat('a', 3);
-        assertThat(shortestExample, equalTo(expected));
-    }
-
-    @Test
-    public void getShortestExample_withComplexRegexContainingOptionalPaths_shouldReturnCorrectString(){
-        Automaton automaton = getAutomaton("[a-z]{0,10}([a-c]{3}|[d-e]{5})");
-
-        String shortestExample = AutomatonUtils.getShortestExample(automaton);
-
-        assertThat(shortestExample, matchesPattern("[a-z]{3}"));
-    }
-
-    @Test
-    public void getShortestExample_withLengthLimitedEmailAddress_shouldReturnCorrectStringWithLowestOrdinalCharPerSection(){
-        Automaton email = getAutomaton("^[a-z0-9]+\\@[a-z0-9]+\\.co(m|\\.uk)$");
-        Automaton lengthRestriction = getAutomaton("^.{0,20}$");
-        Automaton intersected = email.intersection(lengthRestriction);
-
-        String longestExample = AutomatonUtils.getShortestExample(intersected);
-
-        assertThat(longestExample.length(), equalTo(7));
-        assertThat(longestExample, matchesPattern("^[a-z0-9]+\\@[a-z0-9]+\\.co(m|\\.uk)$"));
-    }
-
-    @Test
-    public void getShortestExample_withEmptyString_shouldReturnEmptyString(){
-        Automaton automaton = getAutomaton("[a-z]{0}");
-
-        String longestExample = AutomatonUtils.getShortestExample(automaton);
-
-        assertThat(longestExample, equalTo(""));
-    }
 
     @Test
     public void createAutomaton_withValidString_shouldAcceptValidCharacters(){
@@ -198,13 +60,5 @@ class AutomatonUtilsTests {
         Automaton automaton = AutomatonUtils.createAutomaton(validRegex, true, dummyCache);
 
         assertTrue(automaton.isEmpty());
-    }
-
-    private static Automaton getAutomaton(String regex){
-        return AutomatonUtils.createAutomaton(regex, true, new HashMap<>());
-    }
-
-    private static String repeat(Character character, int times){
-        return IntStream.range(0, times).mapToObj(index -> character.toString()).collect(Collectors.joining(""));
     }
 }
